use a gulong when remembering the return value of g_signal_connect(). we were using a guint in some places.

+++ gtk/tr-core.c
@@ -70,7 +70,7 @@ struct TrCorePrivate
     gulong         monitor_tag;
     GFile        * monitor_dir;
     GSList       * monitor_files;
-    guint          monitor_idle_tag;
+    gulong         monitor_idle_tag;
 
     gboolean       adding_from_watch_dir;
     gboolean       inhibit_allowed;
